add test helpers to reduce code duplication
